requirements.txt
# Voice-to-Voice RAG with Ultravox
# Complete Requirements

# ================================
# Voice Pipeline Dependencies
# ================================

# WebSocket
websockets>=12.0

# Audio
sounddevice>=0.4.6
numpy>=1.24.0

# HTTP Client
httpx>=0.25.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web Server (for RAG tool endpoint)
fastapi>=0.109.0
uvicorn>=0.27.0

# Logging
loguru>=0.7.0

# ================================
# RAG Dependencies (Existing)
# ================================

# HTTP requests (for Tika document parser)
requests>=2.28.0

# NLP - spaCy for text chunking
spacy>=3.5.0

# Vector Database
pymilvus>=2.3.0

# Embeddings
sentence-transformers>=2.2.0

# PyTorch (required by sentence-transformers)
torch>=2.0.0

# LLM (Groq)
groq>=0.4.0
